+++ tests/unit/v1/test_validate.py
@@ -3,7 +3,12 @@ import pydantic
 import pytest
 from freezegun import freeze_time
 
+from data_inclusion.schema import validation
 from data_inclusion.schema.v1 import Service, Structure
+
+
+def has_error(loc: str, exc: pydantic.ValidationError) -> bool:
+    return any(error["loc"] == (loc,) for error in exc.errors())
 
 
 @freeze_time("2024-01-01")
@@ -37,16 +42,31 @@ def test_date_maj_valide(factory, date_maj, est_valide):
         ("*" * 3, True),
         ("*" * 150, True),
         ("*" * 151, False),
-        ("Centre social Le Tournesol.", False),
-        ("Centre social Le Tournesol", True),
-        ("Centre social Le Tournesol etc.", True),
     ],
 )
-def test_nom_valide(factory, nom, est_valide):
+def test_longueur_nom(model, nom, est_valide):
     try:
-        factory(nom=nom)
+        model.model_validate({"nom": nom})
     except pydantic.ValidationError as exc:
-        assert not est_valide
-        assert "nom" in str(exc)
+        assert est_valide != has_error("nom", exc)
+    else:
+        assert est_valide
+
+
+@pytest.mark.parametrize("model", [Structure, Service])
+@pytest.mark.parametrize(
+    ("nom", "mode", "est_valide"),
+    [
+        ("Centre social Le Tournesol.", validation.Mode.NORMAL, True),
+        ("Centre social Le Tournesol.", validation.Mode.STRICT, False),
+        ("Centre social Le Tournesol", validation.Mode.NORMAL, True),
+        ("Centre social Le Tournesol etc.", validation.Mode.NORMAL, True),
+    ],
+)
+def test_nom_ne_se_termine_pas_par_un_point(model, mode, nom, est_valide):
+    try:
+        model.model_validate({"nom": nom}, context={"mode": mode})
+    except pydantic.ValidationError as exc:
+        assert est_valide != has_error("nom", exc)
     else:
         assert est_valide